Remove move feature
