+++ Cargo.toml
@@ -34,9 +34,6 @@ members = [
     "programs/exchange_api",
     "programs/exchange_program",
     "programs/failure_program",
-    "programs/move_loader_api",
-    "programs/move_loader_program",
-    "programs/librapay_api",
     "programs/noop_program",
     "programs/stake_api",
     "programs/stake_program",
@@ -61,4 +58,7 @@ members = [
 
 exclude = [
     "programs/bpf",
+    "programs/move_loader_api",
+    "programs/move_loader_program",
+    "programs/librapay_api",
 ]
