+++ bench-tps/Cargo.toml
@@ -35,4 +35,4 @@ serial_test = "0.2.0"
 serial_test_derive = "0.2.0"
 
 [features]
-move = ["solana-core/move", "solana-librapay-api", "solana-move-loader-program", "solana-move-loader-api"]
+move = ["solana-librapay-api", "solana-move-loader-program", "solana-move-loader-api"]
