+++ core/Cargo.toml
@@ -15,7 +15,6 @@ codecov = { repository = "solana-labs/solana", branch = "master", service = "git
 
 [features]
 pin_gpu_memory = []
-move = []
 
 [dependencies]
 bincode = "1.2.0"
