+++ genesis_programs/Cargo.toml
@@ -18,8 +18,6 @@ solana-config-api = { path = "../programs/config_api", version = "0.21.0" }
 solana-config-program = { path = "../programs/config_program", version = "0.21.0" }
 solana-exchange-api = { path = "../programs/exchange_api", version = "0.21.0" }
 solana-exchange-program = { path = "../programs/exchange_program", version = "0.21.0" }
-solana-move-loader-program = { path = "../programs/move_loader_program", version = "0.21.0", optional = true }
-solana-move-loader-api = { path = "../programs/move_loader_api", version = "0.21.0", optional = true }
 solana-runtime = { path = "../runtime", version = "0.21.0" }
 solana-sdk = { path = "../sdk", version = "0.21.0" }
 solana-stake-api = { path = "../programs/stake_api", version = "0.21.0" }
@@ -34,6 +32,3 @@ solana-vote-program = { path = "../programs/vote_program", version = "0.21.0" }
 [lib]
 crate-type = ["lib"]
 name = "solana_genesis_programs"
-
-[features]
-move = ["solana-move-loader-program", "solana-move-loader-api"]
