+++ programs/move_loader_program/src/lib.rs
@@ -1,12 +1,2 @@
-#[macro_export]
-macro_rules! solana_move_loader_program {
-    () => {
-        (
-            "solana_move_loader_program".to_string(),
-            solana_move_loader_api::id(),
-        )
-    };
-}
-
 use solana_move_loader_api::processor::process_instruction;
 solana_sdk::solana_entrypoint!(process_instruction);
